[None][test] shared KV-transfer harness, MiniMax M3 suite and l0_h100 entries
Extract the shared KV-transfer harness from the DSv4 test module and
port DSv4 onto it; add the MiniMax M3 KV-transfer suite (per-class pool
views, coalesced vs separate schemes) and DSA indexer-K transceiver
coverage; register the disagg KV-transfer suites in the l0_h100 test
list.
