+++ tests/integration/test_lists/test-db/l0_h100.yml
@@ -86,6 +86,9 @@ l0_h100:
   - unittest/disaggregated/test_request_id.py
   - unittest/disaggregated/test_kv_transfer.py
   - unittest/disaggregated/test_kv_transfer_mp.py
+  - unittest/disaggregated/test_transceiver_bounded_polling.py
+  - unittest/disaggregated/test_pool_matching.py
+  - unittest/disaggregated/test_deepseek_v4_kv_transfer.py
   - unittest/disaggregated/test_minimax_m3_kv_transfer.py
   # Split the large cache transceiver parametrization into readable chunks.
   # Main transfer matrix.
@@ -101,6 +104,8 @@ l0_h100:
   # Boundary request lengths.
   - unittest/disaggregated/test_cache_transceiver_single_process.py::test_cache_transceiver_boundary_lengths -k "v1"
   - unittest/disaggregated/test_cache_transceiver_single_process.py::test_cache_transceiver_boundary_lengths -k "v2"
+  # DSA indexer-K side cache (V1, REPLICATED).
+  - unittest/disaggregated/test_cache_transceiver_single_process.py::test_cache_transceiver_v1_dsa_indexer
   - unittest/disaggregated/test_cache_transceiver_harness_report.py
   - unittest/disaggregated/test_cache_transceiver_harness.py
   - unittest/others/test_kv_cache_transceiver.py::test_kv_cache_transceiver_single_process[PYTHON-mha-ctx_fp16_gen_fp16]
