+++ tests/unittest/disaggregated/test_minimax_m3_kv_transfer.py
@@ -1,18 +1,34 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 """MiniMax M3 heterogeneous-topology KV transfer tests.
 
-This reuses the threaded NIXL harness from the DeepSeek-V4 V2 transfer test,
-but supplies MiniMax M3 cache managers and validates both ordinary K/V and the
-replicated INDEX_KEY cache. The TP=2 TEP layout coalesces K/V/INDEX_KEY in one
-physical pool, while attention-DP keeps INDEX_KEY in a separate pool.
+This drives the shared threaded NIXL harness (``kv_transfer_harness``) with
+MiniMax M3 cache managers and validates both ordinary K/V and the
+replicated INDEX_KEY cache. V2 storage coalesces buffers purely by (life
+cycle, size), so at TP degrees where K == V == INDEX_KEY bytes per block the
+index-K cache shares the K/V pool and the slot interleaves per layer; at
+other degrees it gets its own pool. The disagg page-table builder splits each
+pool into per-mapper-kind views, so both layouts (and transfers between them)
+are exercised by the topology matrix below.
 """
 
 from collections.abc import Sequence
 
+import kv_transfer_harness as transfer_harness
 import pytest
-import test_deepseek_v4_kv_transfer as transfer_harness
 import torch
 
 from tensorrt_llm import Mapping
@@ -71,7 +87,10 @@ def test_minimax_cache_indices_support_block_count_limit() -> None:
 def test_v2_disagg_role_mapper_kind_defaults() -> None:
     manager = object.__new__(KVCacheManagerV2)
 
-    assert manager.get_disagg_role_mapper_kinds() == {Role.ALL: MapperKind.INDEXED}
+    assert manager.get_disagg_role_mapper_kinds() == {
+        Role.ALL: MapperKind.INDEXED,
+        Role.INDEX_KEY: MapperKind.REPLICATED,
+    }
 
 
 def test_minimax_disagg_role_mapper_kinds() -> None:
@@ -102,7 +121,9 @@ def test_minimax_disagg_rejects_unmanaged_index_value(monkeypatch) -> None:
         )
 
 
-def _create_manager(mapping: Mapping, dtype: DataType) -> MiniMaxM3KVCacheManagerV2:
+def _create_manager(
+    mapping: Mapping, dtype: DataType, sparse_layers: list[int] | None = None
+) -> MiniMaxM3KVCacheManagerV2:
     max_num_tokens = 2048
     kv_cache_dtype = {
         DataType.FP8: "fp8",
@@ -126,18 +147,72 @@ def _create_manager(mapping: Mapping, dtype: DataType) -> MiniMaxM3KVCacheManage
         dtype=dtype,
         vocab_size=transfer_harness.VOCAB_SIZE,
         max_num_tokens=max_num_tokens,
-        sparse_layer_ids=SPARSE_LAYERS,
-        disable_index_value_layer_ids=SPARSE_LAYERS,
+        sparse_layer_ids=sparse_layers if sparse_layers is not None else SPARSE_LAYERS,
+        disable_index_value_layer_ids=sparse_layers if sparse_layers is not None else SPARSE_LAYERS,
         sparse_index_dim=INDEX_DIM,
     )
+
+
+def test_minimax_m3_pool_view_scheme_coalesced_vs_separate() -> None:
+    """Pool composition varies with TP; the view scheme must not.
+
+    TP=2 makes kv_heads_per_rank == 1, so K == V == INDEX_KEY bytes per
+    block and V2 coalesces all three into one pool whose slot interleaves
+    the sparse layer's index-K between K/V regions (non-uniform layer
+    stride). TP=1 doubles K/V, so index-K gets its own pool and strides are
+    uniform. Both layouts must yield the same two per-class views.
+    """
+    from tensorrt_llm._torch.disaggregation.resource.kv_extractor import (
+        build_page_table_from_manager,
+    )
+    from tensorrt_llm._torch.disaggregation.resource.utils import get_layer_byte_ranges
+
+    # --- TP=2: coalesced, interleaved slot ---
+    manager = _create_manager(
+        Mapping(world_size=2, rank=0, tp_size=2, pp_size=1), DataType.BF16, sparse_layers=[1]
+    )
+    try:
+        lg = build_page_table_from_manager(manager).layer_groups[0]
+        assert len(lg.pool_views) == 2
+        kv_view = next(pv for pv in lg.pool_views if pv.mapper_kind == MapperKind.NHD)
+        idx_view = next(pv for pv in lg.pool_views if pv.mapper_kind == MapperKind.REPLICATED)
+        assert kv_view.pool_role == frozenset({str(Role.KEY), str(Role.VALUE)})
+        assert idx_view.pool_role == frozenset({str(Role.INDEX_KEY)})
+        # Coalesced: both views address the same physical pool.
+        assert kv_view.pool_idx == idx_view.pool_idx
+        starts, bytes_per_layer = get_layer_byte_ranges(kv_view)
+        unit = bytes_per_layer // 2  # one K or V buffer per block
+        # Slot: L0K L0V L1K L1V L1IDX L2K L2V L3K L3V — the sparse layer's
+        # index-K makes the K/V layer stride non-uniform.
+        assert starts == {0: 0, 1: 2 * unit, 2: 5 * unit, 3: 7 * unit}
+        idx_starts, idx_bytes_per_layer = get_layer_byte_ranges(idx_view)
+        assert idx_starts == {1: 4 * unit}
+        assert idx_bytes_per_layer == unit
+    finally:
+        manager.shutdown()
+
+    # --- TP=1: separate pools, uniform strides ---
+    manager = _create_manager(
+        Mapping(world_size=1, rank=0, tp_size=1, pp_size=1), DataType.BF16, sparse_layers=[1]
+    )
+    try:
+        lg = build_page_table_from_manager(manager).layer_groups[0]
+        assert len(lg.pool_views) == 2
+        kv_view = next(pv for pv in lg.pool_views if pv.mapper_kind == MapperKind.NHD)
+        idx_view = next(pv for pv in lg.pool_views if pv.mapper_kind == MapperKind.REPLICATED)
+        assert kv_view.pool_idx != idx_view.pool_idx
+        starts, bytes_per_layer = get_layer_byte_ranges(kv_view)
+        assert starts == {i: i * bytes_per_layer for i in range(NUM_LAYERS)}
+    finally:
+        manager.shutdown()
 
 
 def _create_managers(
     tp: int,
     pp: int,
     enable_dp: bool,
-    _unused_layout: list[int],
     dtype: DataType = DataType.BF16,
+    sparse_layers: list[int] | None = None,
 ) -> list[MiniMaxM3KVCacheManagerV2]:
     return [
         _create_manager(
@@ -149,6 +224,7 @@ def _create_managers(
                 enable_attention_dp=enable_dp,
             ),
             dtype,
+            sparse_layers,
         )
         for rank in range(tp * pp)
     ]
@@ -177,10 +253,15 @@ def _get_nvfp4_scale_view(
     scale_roles = frozenset({"key_block_scale", "value_block_scale"})
     for layer_group_id, layer_group in enumerate(page_table.layer_groups):
         for pool_view in getattr(layer_group, "pool_views", ()):
+            # Scale buffers land in their own pool (smaller size class), so
+            # selecting the pool by role set suffices; every entry for the
+            # layer inside that pool is a scale buffer.
+            if not (pool_view.pool_role & scale_roles):
+                continue
             matching_entries = [
                 entry
-                for entry, role in zip(pool_view.buffer_entries, pool_view.buffer_roles)
-                if int(entry["local_layer_id"]) == local_layer_id and role in scale_roles
+                for entry in pool_view.buffer_entries
+                if int(entry["local_layer_id"]) == local_layer_id
             ]
             if not matching_entries:
                 continue
@@ -322,7 +403,6 @@ def _initialize_cache(
 
 def _verify_cache(
     request_lengths: list[int],
-    compress_ratios: list[int],
     ctx_managers: Sequence[MiniMaxM3KVCacheManagerV2],
     gen_managers: Sequence[MiniMaxM3KVCacheManagerV2],
     ctx_tp: int,
@@ -334,7 +414,7 @@ def _verify_cache(
     ctx_request_ids: list[int],
     gen_request_ids: list[int],
 ) -> None:
-    del compress_ratios, ctx_pp
+    del ctx_pp
 
     for req_idx, _request_length in enumerate(request_lengths):
         gen_request_id = gen_request_ids[req_idx]
@@ -481,17 +561,56 @@ def test_minimax_m3_kv_transfer(
     cache_dtype,
     update_before_transfer,
 ) -> None:
-    transfer_harness.run_deepseek_v4_transfer_test(
+    transfer_harness.run_kv_transfer_test(
         ctx_tp=ctx_tp,
         ctx_pp=ctx_pp,
         gen_tp=gen_tp,
         gen_pp=gen_pp,
         ctx_enable_dp=ctx_enable_dp,
         gen_enable_dp=gen_enable_dp,
-        compress_ratios=[1] * NUM_LAYERS,
         update_before_transfer=update_before_transfer,
-        manager_factory=lambda tp, pp, enable_dp, layout: _create_managers(
-            tp, pp, enable_dp, layout, cache_dtype
+        manager_factory=lambda tp, pp, enable_dp: _create_managers(tp, pp, enable_dp, cache_dtype),
+        init_fn=_initialize_cache,
+        verify_fn=_verify_cache,
+    )
+
+
+# Multiple sparse layers spread across PP stages: the replicated index-key
+# pool overlaps only partially between peers, exercising the layer-strided
+# ReplicatedMapper offsets (a single-sparse-layer model always fully
+# overlaps and would degenerate to a whole-slot copy).
+MULTI_SPARSE_LAYERS = [1, 2, 3]
+
+
+@pytest.mark.cuda
+@pytest.mark.timeout(180)
+@pytest.mark.parametrize(
+    "ctx_tp,ctx_pp,ctx_enable_dp,gen_tp,gen_pp,gen_enable_dp",
+    [
+        (1, 1, True, 1, 2, False),  # ctx full index pool -> per-PP-stage subsets
+        (1, 2, False, 1, 1, True),  # per-PP-stage subsets -> full index pool
+        (2, 1, False, 1, 2, False),  # TEP fan-in + PP subset on the same transfer
+    ],
+    ids=["dep1_to_tp1pp2", "tp1pp2_to_dep1", "tep2_to_tp1pp2"],
+)
+def test_minimax_m3_multi_sparse_layer_pp_transfer(
+    ctx_tp,
+    ctx_pp,
+    ctx_enable_dp,
+    gen_tp,
+    gen_pp,
+    gen_enable_dp,
+) -> None:
+    transfer_harness.run_kv_transfer_test(
+        ctx_tp=ctx_tp,
+        ctx_pp=ctx_pp,
+        gen_tp=gen_tp,
+        gen_pp=gen_pp,
+        ctx_enable_dp=ctx_enable_dp,
+        gen_enable_dp=gen_enable_dp,
+        update_before_transfer=True,
+        manager_factory=lambda tp, pp, enable_dp: _create_managers(
+            tp, pp, enable_dp, DataType.BF16, sparse_layers=MULTI_SPARSE_LAYERS
         ),
         init_fn=_initialize_cache,
         verify_fn=_verify_cache,
